# openapi.yaml
openapi: 3.1.0
info:
  title: Agriculture Masterdata Sync API
  version: 0.1.0-alpha1
  description: >
    HTTP surface for the Agriculture Masterdata Sync protocol (AgmaSync). It lets a
    connected participant send, deactivate, request, and read the master-data
    entities (organizations, persons, farms, fields, field boundaries) that
    agrirouter synchronizes across an n:m network of systems, and configure
    per-entity opt-in for an endpoint.

    Each entity type has its own set of paths and its own concrete schema; there
    is no generic entity resource. The union type `Entity` survives only where a
    single channel genuinely carries all types — the event stream.

    agrirouter holds the Single Source of Truth (SSOT): the canonical copy of
    every entity plus a mapping to each participant's own identifier.

    A write answers with the resulting canonical object, and that response is a
    delivery rather than an acknowledgement: it is the only place the writing
    endpoint sees the revision it just produced, since agrirouter does not echo a
    change back to its source. Clients MUST apply it as they apply an event from
    the stream, and MUST NOT apply any object whose `revision` is lower than the
    one they already hold. See "Applying what agrirouter returns" in
    specification.md.

    Attributes whose value set is expected to grow follow the Zalando RESTful
    API Guidelines convention for open-ended value lists
...
  contact:
...
security:
  - bearerAuth: []
...
paths:
...
  /masterdata/events:
    get:
      tags: [Events]
      summary: Receive master-data changes (Server-Sent Events)
      description: >
        A persistent SSE stream of master-data changes for the calling
        application, carrying every tenant it is routed to and every entity type
        it is opted into. agrirouter never echoes a change back to the endpoint
        it originated from; a change made by one endpoint is still delivered to
        the application's other endpoints. Each event's
        `data` is an `Entity` — an `Organization`, `Person`, `Farm`, `Field`, or
        `FieldBoundary` discriminated by its `type` property. A delivered object
        carries the receiving application's own identifier in `local_id` when
        agrirouter holds one, and no `local_id` at all when it does not.

        This stream carries steady-state synchronization only. Initial load is
        delivered separately, per endpoint, by
        `/endpoints/{external_id}/masterdata-initial-load/events`;
        the two are independent and are not deduplicated against each other, so an
        object may arrive on both while an endpoint is loading.

        An application reconnecting after an absence is served catch-up before
        live changes: everything it is entitled to that changed since its
        position, ordered so that a referenced object precedes the objects that
        reference it, ending with a `CAUGHT_UP` event. That is the only property
        of the order an application may rely on. The order itself is
        agrirouter's and may change in a later version of this API, so an
        application MUST NOT depend on the position of one entity type relative
        to another, or read the completeness of an entity type out of it.
        `Last-Event-ID` is the only position the application keeps, and
        positions do not expire.

        The stream also carries `ROUTE_CHANGED`, which states which
        entity types the user has selected for one of the application's
        endpoints. It is issued every time that selection changes — the endpoint
        routed to master-data, a type selected, a type deselected, the last one
        deselected.

        It also carries `RESET_MASTERDATA_SYNC` when the user has wiped a
        tenant's master data: every canonical object, identifier mapping,
        masterdata route and initial-load state of the tenant is gone, and the
        application MUST discard its bindings there. It reaches every
        application holding a binding in the tenant, including one with no
        endpoint left there, so it may concern a tenant the application is no
        longer routed to. It precedes any later event of that tenant, catch-up
        included. The application MUST durably
        store a position at or past it before taking part in that tenant's
        initial load again.

      operationId: streamMasterdataEvents
      parameters:
        - $ref: '#/components/parameters/LastEventId'
      responses:
        '200':
          description: An event stream.
          content:
            text/event-stream:
              schema:
                description: >
                  SSE stream as defined in https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events

                  SSE event would carry three parts: "id:", "event:" and "data:".

                  The `data:` is a payload of each SSE frame. Frames carry an `event:`
                  of `MASTERDATA_CHANGED`, `MASTERDATA_DEACTIVATED`,
                  `CAUGHT_UP`, `ROUTE_CHANGED` or `RESET_MASTERDATA_SYNC`; the
                  first two carry an `Entity`, `CAUGHT_UP` has no data, and
                  `ROUTE_CHANGED` and `RESET_MASTERDATA_SYNC` carry an event payload.
                  Every frame also carries an `id:`, which is the
                  delivery position as of that frame — what the application
                  persists once it has durably applied the frame, and returns as
                  `Last-Event-ID` on reconnect. `id` may repeat across some events
                  and application should not expect it would always advance or
                  attempt to parse, compare, or construct it in any way: `id` must
                  be treated as an opaque string whose structure may change.
                  Whatever `id` a frame carries is safe to resume from once that
                  frame and every frame before it have been applied.
                oneOf:
                  - $ref: '#/components/schemas/Entity'
                  - $ref: '#/components/schemas/RouteChangedEventData'
                  - $ref: '#/components/schemas/MasterdataResetEventData'

  # ------------------------------------------------------------ configuration
  /endpoints/{external_id}:
    put:
      tags: [Configuration]
      operationId: putEndpoint
      summary: Create or update endpoint
      description: >
        This resource can be used to create new endpoint or reconfigure existing one.

        
        Masterdata: Declare what masterdata entity types the endpoint can exchange. 
        Masterdata routes must be separately created by a user, selecting a subset of 
        these entity types. There is no default routing for masterdata. 
        The configuration MUST be dependency-closed: enabling fields requires
        the farms and field boundaries those fields reference, and the
        organizations and persons those farms reference, to be enabled as well. A configuration that is
        not dependency-closed is rejected with `400`.
      security:
        - agrirouterOauthQA: ["endpoints:manage"]
        - agrirouterOauthPROD: ["endpoints:manage"]
      parameters:
        - $ref: "#/components/parameters/AgrirouterTenantId"
        - $ref: "#/components/parameters/external_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutEndpointRequest"
      responses:
        '200':
          description: Endpoint was successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        '201':
          description: Endpoint was successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        '400':
          description: Request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Authorization failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '413':
          description: Content too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '502':
          description: Bad Gateway, upstream service has returned an error
        '503':
          description: Service Unavailable at the moment
        '504':
          description: Gateway Timeout, request took too long to process
        '500':
          description: Server Side Error

  # ----------------------------------------------------------------- initial load
  /endpoints/{external_id}/masterdata-initial-load/events:
    parameters:
      - $ref: '#/components/parameters/AgrirouterTenantId'
      - $ref: '#/components/parameters/external_id'
    get:
      tags: [InitialLoad]
      summary: Receive the endpoint's canonical set (Server-Sent Events)
      description: >
        Delivers the canonical set this endpoint is owed: every object of every
        entity type it is opted into that it is entitled to, while the endpoint
        is in `LOADING_FROM_AGRIROUTER`. One stream per endpoint, matching the
        granularity of the initial-load state and the shape of its status
        resource. The set is complete: objects that are deactivated are part of
        it and arrive with `active: false`, and so are objects whose current
        revision the requesting application itself last wrote — origin
        suppression does not apply here, since withholding them would have a
        returning endpoint push them back as new and duplicate them. A delivered
        object carries the application's own identifier in `local_id` where
        agrirouter holds one.

        agrirouter moves the endpoint to `RECONCILING` once it has sent the
        whole set and closes the HTTP response after it, in that order: the
        state is what tells an orderly finish from a dropped connection, so
        closing first would leave a window in which an endpoint reads
        `LOADING_FROM_AGRIROUTER` for a set that did arrive and takes the whole
        set again for nothing. The set is fixed when the load
        starts: an entity type opted in while the set is streaming restarts the
        load, so the endpoint re-enters `LOADING_FROM_AGRIROUTER` and the set is
        sent again including the new type.

        Order is agrirouter's, not the endpoint's. The set is delivered so that a
        referenced object precedes the objects that reference it, as catch-up on
        `/masterdata/events` is, and opt-in is dependency-closed, so the target
        of every reference is in the set and an endpoint can apply each object
        as it arrives. That is the only property of the order an endpoint may
        rely on: the order itself may change in a later version of this API, so
        an endpoint MUST NOT depend on the position of one entity type relative
        to another, or read the completeness of an entity type out of it. An
        object referenced from the live stream that this stream has not
        delivered yet is requested (`POST /masterdata/<types>/requests`).

        The stream carries **no delivery position**: it delivers a set rather than
        a sequence of changes, so it takes no `Last-Event-ID`, its frames carry no
        `id:`, and a connection that drops before the set is complete is recovered
        by requesting it again from the beginning.

        An endpoint MUST NOT treat the response ending as proof that the set
        arrived — a dropped connection ends it the same way. Whether the set was
        sent is recorded in the endpoint's state on the Initial Load resource,
        which agrirouter advances only after sending everything, and before
        closing this response. That read is therefore conclusive as soon as the
        response ends: an endpoint that finds `LOADING_FROM_AGRIROUTER` there
        has been cut short and reconnects at once, without waiting before the
        read or polling for a state change.

        This stream is independent of the application's steady-state stream
        `/masterdata/events`, which keeps running throughout and is not
        deduplicated against this one.
      operationId: streamInitialLoadEvents
      responses:
        '200':
          description: >
            The canonical set, as an event stream. The response closing is what
            marks the set complete.
          content:
            text/event-stream:
              schema:
                description: >
                  SSE stream as defined in https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events

                  SSE event would carry three parts: "id:", "event:" and "data:".

                  Every frame carries an `event:` of `MASTERDATA_CHANGED`
                  and an `Entity` type in the `data:` payload.
                $ref: '#/components/schemas/Entity'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /endpoints/{external_id}/masterdata-initial-load/user-attention:
    parameters:
      - $ref: '#/components/parameters/AgrirouterTenantId'
      - $ref: '#/components/parameters/external_id'
    put:
      tags: [InitialLoad]
      summary: Report that the endpoint's initial load is waiting on a user
      description: >
        Raises `awaiting_user` on the endpoint's initial-load status, so that
        agrirouter shows "waiting for you in <app>" — linking to the master-data
        resolution URI supplied for this endpoint — in place of its own "this
        application is working through your data".


        It is a resource of its own because reporting a user is not a
        transition. Conflicts surface object by object, so an endpoint has to be
        able to say this from any state before `COMPLETED`, including while the
        canonical set is still arriving; naming a state in order to say it would
        mean naming one the endpoint does not own and cannot set. This names
        none, so it never races the transitions agrirouter drives and is never
        out of order.


        It carries no body and raises the flag; there is nothing to lower it
        with, because lowering it is not the endpoint's to do. agrirouter clears
        it on the two endpoint-driven transitions — advancing to
        `LOADING_TO_AGRIROUTER` or to `COMPLETED` — and the step to
        `RECONCILING` does not, that being agrirouter reporting it has finished
        sending rather than the user having finished deciding. Repeating it
        while it is already raised is a `200`.


        agrirouter learns *that* a person is needed and never what for: it is one
        bit per endpoint, not a conflict list. Nothing in the protocol branches
        on it, so an endpoint that never calls this costs precision rather than
        correctness.
      operationId: reportUserAttention
      responses:
        '200':
          description: The endpoint's initial load status, with the flag raised.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitialLoadStatus'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >
            The endpoint's initial load is `COMPLETED`, and a completed load is
            waiting on nobody. An endpoint opted into no entity type has no
            initial-load state at all and is `404`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      description: OAuth 2.0 bearer tokens obtained via the Client Credentials flow.
      flows:
        clientCredentials:
          tokenUrl: https://api.prod.agrirouter.farm/oauth/token
          scopes:
            manage_endpoints: Configure endpoints and their master-data opt-in.
            masterdata: Send, request, and read synchronized master data.
  parameters:
    AgrirouterTenantId:
      name: x-agrirouter-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The farmer's tenant ID in relation to which communication is done.
...
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      description: >
        The event id the application saved last from `/masterdata/events`
        stream, as agrirouter issued it in the `id:` field of an SSE
        frame.
        The application MUST send it back exactly as issued, treat it as an opaque string
        and MUST NOT interpret, compare, construct, or modify it in any way. Its
        structure is not defined by this API and may change; the example below
        illustrates its size and shape only.

        It does not always advance on every frame: the same value may repeat
        on some frames, and an application MUST NOT read anything into that
        beyond its own position for `Last-Event-ID` being unchanged.


        The application MUST save it from what it has **durably applied**
        rather than from whatever its stream client last read; delivery is
        at-least-once, so everything after it is sent again on reconnect. Where
        applies complete out of order, store the id of the earliest frame, in
        the order the stream delivered them, that is not yet applied: ids are
        not comparable by the application, so this is a matter of remembering
        delivery order. It does not expire, and agrirouter stores no copy of it.


        Omitted, or present and one agrirouter cannot validate, is served as a
        first connection on this stream: everything the application is entitled
        to, ending with a `CAUGHT_UP` frame. That costs the application its place
        in the stream and not its entitlement, and is how it asks for every
        entity to be delivered again.


        This parameter applies to `/masterdata/events` only. The initial-load
        stream carries no position and does not accept it.
      schema:
        type: string
      example: 'v1.f8yQ_KL5hHBS.TBQjL_8GeXSs6TIxFzsdQQ'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A human-readable error message describing what went wrong.
...